Restructured file to pages directory
