+++ my-app/src/pages/HomePage.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { useState } from 'react';
-import SectionTitle from './components/SectionTitle/index.jsx';
-import SongItem from './components/SongItem/index.jsx';
-import LoginButton from './components/LoginButton/index.jsx';
-import SearchBar from './components/SearchBar/index.jsx';
-import CreatePlaylistForm from './components/CreatePlaylistForm/index.jsx';
+import SectionTitle from '../components/SectionTitle/index.jsx';
+import SongItem from '../components/SongItem/index.jsx';
+import LoginButton from '../components/LoginButton/index.jsx';
+import SearchBar from '../components/SearchBar/index.jsx';
+import CreatePlaylistForm from '../components/CreatePlaylistForm/index.jsx';
 import axios from 'axios';
 import { useDispatch, useSelector } from 'react-redux';
-import { getToken } from './redux/token.js';
+import { getToken } from '../redux/token.js';
 
 const HomePage = () => {
